requirements.txt
pandas
genson
pymongo
requests
fastapi
uvicorn
streamlit
schedule
python-dotenv
python-multipart
beautifulsoup4
